revert decrease res and change disp to data again

+++ index.php
@@ -427,13 +427,13 @@ $badge = count($_SESSION['cart']);
             $filestr = explode("." ,$item['image']);
             $filepath = "./image-data/" . $filestr[0] . ".txt";
 
-            //if(file_exists($filepath)){
-            //    $file = file_get_contents($filepath, true);
-            //    echo ("<img class=\"myImages\" id=\"".$item['itemid']."\" src=\"".$file." \"loading=\"lazy\">");
-            //}
-            //else{
+            if(file_exists($filepath)){
+                $file = file_get_contents($filepath, true);
+                echo ("<img class=\"myImages\" id=\"".$item['itemid']."\" src=\"".$file." \"loading=\"lazy\">");
+            }
+            else{
                 echo("<img class=\"myImages\" id=\"".$item['itemid']."\" src=\"item-image/".($item['image'])." \"loading=\"lazy\">");
-            //}
+            }
             
             echo("<h3>".$item['name']."</h3>");
             echo("<div class=\"weight-size\">".$item['weight']." gr");
@@ -477,13 +477,13 @@ $badge = count($_SESSION['cart']);
             $filestr = explode("." ,$item['image']);
             $filepath = "./image-data/" . $filestr[0] . ".txt";
 
-            //if(file_exists($filepath)){
-            //    $file = file_get_contents($filepath, true);
-            //    echo ("<img class=\"myImages\" id=\"".$item['itemid']."\" src=\"".$file." \"loading=\"lazy\">");
-            //}
-            //else{
+            if(file_exists($filepath)){
+                $file = file_get_contents($filepath, true);
+                echo ("<img class=\"myImages\" id=\"".$item['itemid']."\" src=\"".$file." \"loading=\"lazy\">");
+            }
+            else{
                 echo("<img class=\"myImages\" id=\"".$item['itemid']."\" src=\"item-image/".($item['image'])." \" loading=\"lazy\">");
-            //}
+            }
             echo("<h3>".$item['name']."</h3>");
             echo("<div class=\"weight-size\">".$item['weight']." gr");
             if($item['size']>0){
